add main card and rename prev card to AppCardEdit
