+++ src/components/AppCard.tsx
@@ -1,29 +1,23 @@
 import React from 'react';
-import style from './AppCard.module.css';
-import binIcon from '../images/Kind=Delete.svg';
-//card
+import styles from './AppCard.module.css';
+import editIcon from '../images/Kind=Edit.svg';
+
 export const AppCard = () => {
-	return (
-		<>
-			<div className={style.cardConteinerNext}>
-				<textarea className={style.inputText}></textarea>
-				<div>
-					<button className={style.btnCancel}>CANCEL</button>
-					<button className={style.btnConfirm}>NEXT</button>
-				</div>
+	const textArr: Array<string> = [
+		'What is Lorem Ipsum?',
+		'Lorem Ipsum is simply dummy text of the printing',
+		'and typesetting industry. Lorem Ipsum has been the industry',
+		'standard dummy text ever since the 1500s, when an unknown printer',
+		'took a galley of type and scrambled it to make a type',
+		'specimen book. It has survived not only five centuries, but also the leap',
+		'into electronic typesetting, remaining essentially unchanged.',
+	];
+	return textArr.map((item, index) => (
+		<div className={styles.card} key={index}>
+			<div className={styles.editConteiner}>
+				<img src={editIcon} alt='edit' />
 			</div>
-			<div className={style.cardConteinerSave}>
-				<div className={style.saveHeader}>
-					<button className={style.binBtn}>
-						<img src={binIcon} alt='binIcon' />
-					</button>
-				</div>
-				<textarea className={style.inputText}></textarea>
-				<div>
-					<button className={style.btnCancel}>BACK</button>
-					<button className={style.btnConfirm}>SAVE</button>
-				</div>
-			</div>
-		</>
-	);
+			<h4 className={styles.text}>{item}</h4>
+		</div>
+	));
 };
