+++ src/components/AppCardEdit.module.css
@@ -1,0 +1,60 @@
+/* card */
+.cardConteinerNext,
+.cardConteinerSave {
+	display: flex;
+	flex-direction: column;
+	margin: 10px;
+	padding: 10px;
+	border: 1px solid;
+	border-radius: 10px;
+}
+.saveHeader {
+	margin-top: 0px;
+}
+/* input */
+.inputText {
+	width: 100%;
+	min-height: '40px';
+	resize: none;
+	border: 1px solid gray;
+}
+/* buttons */
+.cardConteinerNext div,
+.cardConteinerSave div {
+	display: flex;
+	justify-content: right;
+	margin-top: 10px;
+}
+.cardConteinerSave .saveHeader {
+	margin-top: 0px;
+}
+.cardConteinerNext div button,
+.cardConteinerSave div button {
+	padding: 6px 20px;
+	margin-right: 8px;
+	border: 0px;
+	border-radius: 20px;
+}
+.cardConteinerNext div button:first-child,
+.cardConteinerSave div button:first-child {
+	padding: 0 10px;
+}
+.cardConteinerNext div button:last-child,
+.cardConteinerSave div button:last-child {
+	margin-right: 0px;
+}
+.binBtn {
+	padding: -10px 0;
+	border: none;
+}
+.btnCancel {
+	border: none;
+}
+.btnConfirm {
+	color: azure;
+	background-color: black;
+}
+.btnDelete {
+	color: azure;
+	background-color: #d11a2a;
+}
