+++ src/components/AppHeader.module.css
@@ -8,6 +8,7 @@
 	position: fixed;
 	top: 0;
 	width: 100%;
+	border-bottom: 1px solid;
 }
 .logo {
 	width: 31.815px;
