+++ src/components/AppLayout.module.css
@@ -1,6 +1,6 @@
 .layout {
 	max-width: 390px;
-	height: 844px;
+	height: 100%;
 	margin: 1% auto;
 	background: #f1f2f3;
 }
